requirements.txt
Pillow
numpy
colorama
unicodedata2